Publish new CanvasLayer libraries (CanvasLayerContainer, CanvasLayers, and CanvasLayerElements)
